Fixed the Deserialization Issue and Added Eureka Client Registry

+++ spring-jpa/src/main/java/com/devesh/springjpa/service/ReservationService.java
@@ -4,5 +4,5 @@ import com.devesh.springjpa.model.Reservation;
 
 public interface ReservationService {
 
-	public Reservation searchRestaurantID(Integer restID);
+	public Reservation searchRestaurantID(Long restID);
 }
